requirements.txt
apscheduler
requests
python-magic
